moved parsing username and hostname functionality outside of the client class
